Entering and exiting

+++ hibernate_module/src/main/java/il/cshaifasweng/DataManipulationThroughDB/DAO.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface DAO<T> {
     SessionFactory factory =MySQL.getSessionFactory();
-    T get(int id,Class<T> type);
+    T get(long id,Class<T> type);
     T getLastAdded(Class<T> Type);
     List<T> getAll(Class<T> type);
     void save(T t,Class<T> type);
